Oxygen meter gradient
In vector calculus, the gradient of a scalar-valued differentiable function
� f of several variables is the vector field (or vector-valued function) ∇ � \nabla f whose value at a point �
p is the "direction and rate of fastest increase". If the gradient of a function is non-zero at a point � p, the direction of the gradient is the direction in which the function increases most quickly from � p, and the magnitude of the gradient is the rate of increase in that direction, the greatest absolute directional derivative.[1] Further, a point where the gradient is the zero vector is known as a stationary point. The gradient thus plays a fundamental role in optimization theory, where it is used to maximize a function by gradient ascent. In coordinate-free terms, the gradient of a function � (�).